# Package on main thread for small rebuilds