--- a/packages/core/types/index.js
+++ b/packages/core/types/index.js
@@ -5,7 +5,7 @@ import type SourceMap from '@parcel/source-map';
 import type {FileSystem} from '@parcel/fs';
 import type WorkerFarm from '@parcel/workers';
 import type {PackageManager} from '@parcel/package-manager';
-import type InternalBundleGraph from '@parcel/core';
+import type InternalBundleGraph from '../core/src/BundleGraph';
 import type {
   Diagnostic,
   Diagnostifiable,