<!---
Thanks for filing a pull request 😄 ! Before you submit, please read the following:

Search open/closed issues before submitting since someone might have pushed the same thing before!
-->

# ↪️ Pull Request
This change prevents `WriteBundlesRequest` from serializing across all workers with small rebuilds which _should_ improve save/warm build times when small changes that only impact a single bundle are made to the project.

Tested saving a file with a small change in a large project

Before:
Save time | ✨ Built in 154.08s
-- | --
Packaging phase | 13.614s

After:
Save time | ✨ Built in 134.72s 🚀 12.56% faster
-- | --
Packaging phase | 1.948s 🚀 85.70% faster



<!---
Provide a general summary of the pull request here
Please look for any issues that this PR resolves and tag them in the PR.
-->

## ✔️ PR Todo

- [ ] Added/updated unit tests for this change
- [ ] Filled out test instructions (In case there aren't any unit tests)
- [ ] Included links to related issues/PRs