@devongovett 2022-10-31
I've refactored the WorkerFarm in a way that should make this a bit simpler to implement: #8589.